requirements.txt
openai
langchain
streamlit
faiss-cpu
python-dotenv
